added archive deletion

+++ src/main/java/com/fitness/Service/Archive/ArchiveService.java
@@ -9,11 +9,15 @@ import com.fitness.Model.Report.Salary;
 import com.fitness.Model.Work.Employment;
 import com.fitness.Model.Work.Subscription;
 import com.fitness.Service.Person.CustomerService;
+import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonBar;
+import javafx.scene.control.ButtonType;
 
 import java.sql.SQLException;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 public class ArchiveService {
     private final ArchiveDAO archiveDAO = new ArchiveDAO();
@@ -38,7 +42,33 @@ public class ArchiveService {
     }
 
     public void remove(Archive archive) throws SQLException {
-        archiveDAO.remove(archive, true);
+        if(archive == null) return;
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Հաստատում");
+
+        ButtonType delete = new ButtonType("Ջնջել", ButtonBar.ButtonData.OK_DONE);
+        ButtonType back = new ButtonType("Հետ", ButtonBar.ButtonData.CANCEL_CLOSE);
+
+        alert.getButtonTypes().setAll(delete, back);
+
+        if(archive.isRegistration()) {
+            alert.setHeaderText("Ջնջելով գրանցման օրը, ամբողջությամբ կջնջվեն բոլոր այցերը և տվյալ հաճախորդը ");
+            alert.setContentText("Ցանկանու՞մ եք ջնջել գրանցումը");
+        } else {
+            alert.setHeaderText("");
+            alert.setContentText("Ցանկանու՞մ եք ջնջել տվյալ այցը");
+        }
+
+        Optional<ButtonType> result =  alert.showAndWait();
+
+        if(result.isPresent())
+            if(result.get() == delete)
+                if(archive.isRegistration()) {
+                    customerService.remove(archive.getCustomer(), true);
+                } else {
+                    archiveDAO.remove(archive, true);
+                }
+
     }
 
     public List<Archive> getByDateRange(LocalDate startDate, LocalDate endDate) throws SQLException {
@@ -59,12 +89,14 @@ public class ArchiveService {
         return null;
     }
 
-    private int getPrice(Subscription subscription, Employment employment) {
+    private int getPrice(Subscription subscription, Archive archive) {
         /* if user has subscription, get that subscriptions that employment price */
-        if(subscription != null)
-            return subscription.getEmploymentPrice(employment);
-        else if(employment != null)
-            return employment.getPrice();
+        if(archive.isBonus())
+            return 0;
+        else if(subscription != null)
+            return subscription.getEmploymentPrice(archive.getEmployment());
+        else if(archive.getEmployment() != null)
+            return archive.getEmployment().getPrice();
         else
             return 0;
     }
@@ -72,8 +104,8 @@ public class ArchiveService {
     public List<Report> getReports(List<Archive> archives) {
         List<Report> reports = new ArrayList<>();
         for(Archive archive: archives) {
-            /* pass bonus visits and registrations */
-            if(archive.isBonus() || archive.getEmployment() == null && archive.getEmployee() == null) continue;
+            /* pass registrations */
+            if(archive.isRegistration()) continue;
 
             Report report = getReportByEmployment(reports, archive.getEmployment());
             if(report != null) {
@@ -82,14 +114,14 @@ public class ArchiveService {
                     salary.incrementQuantity();
 
                     Subscription subscription = archive.getCustomer().getSubscription();
-                    salary.incrementPrice(getPrice(subscription, archive.getEmployment()));
+                    salary.incrementPrice(getPrice(subscription, archive));
                 } else { /* employee is not in salary */
                     Salary newSalary = new Salary();
 
                     newSalary.setEmployee(archive.getEmployee());
 
                     Subscription subscription = archive.getCustomer().getSubscription();
-                    newSalary.setPrice(getPrice(subscription, archive.getEmployment()));
+                    newSalary.setPrice(getPrice(subscription, archive));
 
                     newSalary.setQuantity(1);
 
@@ -105,7 +137,7 @@ public class ArchiveService {
                 newSalary.setEmployee(archive.getEmployee());
 
                 Subscription subscription = archive.getCustomer().getSubscription();
-                newSalary.setPrice(getPrice(subscription, archive.getEmployment()));
+                newSalary.setPrice(getPrice(subscription, archive));
 
                 newSalary.setQuantity(1);
